+++ examples/minimal/main.tf
@@ -21,7 +21,7 @@ module "databricks" {
   resource_group_name = azurerm_resource_group.databricks-rg.name
   location            = azurerm_resource_group.databricks-rg.location
 
-  virtual_network_id      = azurerm_virtual_network.databricks-vnet.id
+  vnet_id                 = azurerm_virtual_network.databricks-vnet.id
   virtual_network_rg_name = azurerm_resource_group.databricks-rg.name
 
   private_subnet_name = module.subnets.private_subnet_name