+++ modules/azure-databricks-workspace/main.tf
@@ -9,7 +9,7 @@ resource "azurerm_databricks_workspace" "module-databricks" {
     no_public_ip        = false
     private_subnet_name = var.private_subnet_name
     public_subnet_name  = var.public_subnet_name
-    virtual_network_id  = var.virtual_network_id
+    virtual_network_id  = var.vnet_id
   }
 
   tags = var.tags