+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Tamr Azure Databricks module
 
 # v0.5.0 - November 4th 2020
-* Changes input variable `name` to `workspace_name` for clarity
-* Changes input variable `security-group-name-prefix` to `security_group_name_prefix` for naming consistency
+* Renames `name` to `workspace_name` for clarity
+* Renames `security-group-name-prefix` to `security_group_name_prefix` for naming consistency
+* Renames `virtual_network_id` to `vnet_id`
 
 # v0.4.2 - September 8th 2020
 * Remove data blocks that attach to subnets during runtime